Move exec Observation into a separate file and split out some helper functions.
