+++ execute/plugin.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	mapset "github.com/deckarep/golang-set/v2"
-	"golang.org/x/exp/maps"
 
 	"github.com/smartcontractkit/libocr/commontypes"
 	"github.com/smartcontractkit/libocr/offchainreporting2plus/ocr3types"
@@ -21,7 +20,6 @@ import (
 	"github.com/smartcontractkit/chainlink-ccip/execute/internal/gas"
 	"github.com/smartcontractkit/chainlink-ccip/execute/report"
 	"github.com/smartcontractkit/chainlink-ccip/execute/tokendata"
-	typeconv "github.com/smartcontractkit/chainlink-ccip/internal/libs/typeconv"
 	"github.com/smartcontractkit/chainlink-ccip/internal/plugincommon"
 	"github.com/smartcontractkit/chainlink-ccip/internal/plugincommon/discovery"
 	dt "github.com/smartcontractkit/chainlink-ccip/internal/plugincommon/discovery/discoverytypes"
@@ -163,166 +161,6 @@ func getPendingExecutedReports(
 		"groupedCommits", groupedCommits, "count", len(groupedCommits))
 
 	return groupedCommits, nil
-}
-
-// Observation collects data across two phases which happen in separate rounds.
-// These phases happen continuously so that except for the first round, every
-// subsequent round can have a new execution report.
-//
-// Phase 1: Gather commit reports from the destination chain and determine
-// which messages are required to build a valid execution report.
-//
-// Phase 2: Gather messages from the source chains and build the execution
-// report.
-// nolint:gocyclo // todo
-func (p *Plugin) Observation(
-	ctx context.Context, outctx ocr3types.OutcomeContext, _ types.Query,
-) (types.Observation, error) {
-	var err error
-	var previousOutcome exectypes.Outcome
-
-	if outctx.PreviousOutcome != nil {
-		previousOutcome, err = exectypes.DecodeOutcome(outctx.PreviousOutcome)
-		if err != nil {
-			return types.Observation{}, fmt.Errorf("unable to decode previous outcome: %w", err)
-		}
-		p.lggr.Infow("decoded previous outcome", "previousOutcome", previousOutcome)
-	}
-
-	var discoveryObs dt.Observation
-	// discovery processor disabled by setting it to nil.
-	if p.discovery != nil {
-		discoveryObs, err = p.discovery.Observation(ctx, dt.Outcome{}, dt.Query{})
-		if err != nil {
-			p.lggr.Errorw("failed to discover contracts", "err", err)
-		}
-
-		if !p.contractsInitialized {
-			p.lggr.Infow("contracts not initialized, only making discovery observations",
-				"discoveryObs", discoveryObs)
-			return exectypes.Observation{Contracts: discoveryObs}.Encode()
-		}
-	}
-
-	state := previousOutcome.State.Next()
-	p.lggr.Debugw("Execute plugin performing observation", "state", state)
-	switch state {
-	case exectypes.GetCommitReports:
-		fetchFrom := time.Now().Add(-p.offchainCfg.MessageVisibilityInterval.Duration()).UTC()
-
-		// Phase 1: Gather commit reports from the destination chain and determine which messages are required to build
-		//          a valid execution report.
-		supportsDest, err := p.supportsDestChain()
-		if err != nil {
-			return types.Observation{}, fmt.Errorf("unable to determine if the destination chain is supported: %w", err)
-		}
-		if supportsDest {
-			groupedCommits, err := getPendingExecutedReports(ctx, p.ccipReader, p.destChain, fetchFrom, p.lggr)
-			if err != nil {
-				return types.Observation{}, err
-			}
-
-			// TODO: truncate grouped to a maximum observation size?
-			return exectypes.NewObservation(groupedCommits, nil, nil, nil, nil, discoveryObs).Encode()
-		}
-
-		// No observation for non-dest readers.
-		return types.Observation{}, nil
-	case exectypes.GetMessages:
-		// Phase 2: Gather messages from the source chains and build the execution report.
-		messages := make(exectypes.MessageObservations)
-		if len(previousOutcome.PendingCommitReports) == 0 {
-			p.lggr.Debug("TODO: No reports to execute. This is expected after a cold start.")
-			// No reports to execute.
-			// This is expected after a cold start.
-		} else {
-			commitReportCache := make(map[cciptypes.ChainSelector][]exectypes.CommitData)
-			for _, report := range previousOutcome.PendingCommitReports {
-				commitReportCache[report.SourceChain] = append(commitReportCache[report.SourceChain], report)
-			}
-
-			for srcChain, reports := range commitReportCache {
-				if len(reports) == 0 {
-					continue
-				}
-
-				ranges, err := computeRanges(reports)
-				if err != nil {
-					return types.Observation{}, err
-				}
-
-				// Read messages for each range.
-				for _, seqRange := range ranges {
-					// TODO: check if srcChain is supported.
-					msgs, err := p.ccipReader.MsgsBetweenSeqNums(ctx, srcChain, seqRange)
-					if err != nil {
-						return nil, err
-					}
-					for _, msg := range msgs {
-						if _, ok := messages[srcChain]; !ok {
-							messages[srcChain] = make(map[cciptypes.SeqNum]cciptypes.Message)
-						}
-						messages[srcChain][msg.Header.SequenceNumber] = msg
-					}
-				}
-			}
-		}
-
-		// Regroup the commit reports back into the observation format.
-		// TODO: use same format for Observation and Outcome.
-		groupedCommits := make(exectypes.CommitObservations)
-		for _, report := range previousOutcome.PendingCommitReports {
-			if _, ok := groupedCommits[report.SourceChain]; !ok {
-				groupedCommits[report.SourceChain] = []exectypes.CommitData{}
-			}
-			groupedCommits[report.SourceChain] = append(groupedCommits[report.SourceChain], report)
-		}
-
-		tkData, err1 := p.tokenDataObserver.Observe(ctx, messages)
-		if err1 != nil {
-			return types.Observation{}, fmt.Errorf("unable to process token data %w", err1)
-		}
-
-		costlyMessages, err := p.costlyMessageObserver.Observe(ctx, messages)
-		if err != nil {
-			return types.Observation{}, fmt.Errorf("unable to observe costly messages %w", err)
-		}
-
-		return exectypes.NewObservation(
-			groupedCommits, messages, costlyMessages, tkData, nil, discoveryObs).Encode()
-
-	case exectypes.Filter:
-		// Phase 3: observe nonce for each unique source/sender pair.
-		nonceRequestArgs := make(map[cciptypes.ChainSelector]map[string]struct{})
-
-		// Collect unique senders.
-		for _, commitReport := range previousOutcome.PendingCommitReports {
-			if _, ok := nonceRequestArgs[commitReport.SourceChain]; !ok {
-				nonceRequestArgs[commitReport.SourceChain] = make(map[string]struct{})
-			}
-
-			for _, msg := range commitReport.Messages {
-				sender := typeconv.AddressBytesToString(msg.Sender[:], uint64(p.destChain))
-				nonceRequestArgs[commitReport.SourceChain][sender] = struct{}{}
-			}
-		}
-
-		// Read args from chain.
-		nonceObservations := make(exectypes.NonceObservations)
-		for srcChain, addrSet := range nonceRequestArgs {
-			// TODO: check if srcSelector is supported.
-			addrs := maps.Keys(addrSet)
-			nonces, err := p.ccipReader.Nonces(ctx, srcChain, p.destChain, addrs)
-			if err != nil {
-				return types.Observation{}, fmt.Errorf("unable to get nonces: %w", err)
-			}
-			nonceObservations[srcChain] = nonces
-		}
-
-		return exectypes.NewObservation(nil, nil, nil, nil, nonceObservations, discoveryObs).Encode()
-	default:
-		return types.Observation{}, fmt.Errorf("unknown state")
-	}
 }
 
 func (p *Plugin) ValidateObservation(
